workspace/pipelines/template/requirements.txt
azureml-core==1.22.0